requirements.txt
2to3==1.0
aiohttp==3.6.2
aiosqlite==0.15.0
appier==1.19.6
arrow==0.15.8
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
backcall==0.1.0
blessings==1.7
bpython==0.19
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
curtsies==0.3.4
databases==0.3.2
decorator==4.4.2
discord-ext-menus==1.0.0a26+g84caae8
discord.py==1.3.4
greenlet==0.4.16
ics==0.7
idna==2.9
install==1.3.3
ipdb==0.13.1
ipython==7.13.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.16.0
joblib==0.16.0
lxml==4.5.0
multidict==4.7.5
nltk==3.5
numpy==1.19.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
PyPika==0.38.0
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==3.1.1
six==1.14.0
SQLAlchemy==1.3.18
tabulate==0.8.7
TatSu==4.4.0
tortoise-orm==0.16.14
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
websockets==8.1
yarl==1.4.2
